+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  
 ### Added
 - Singapore - Add Hari Raya Puasa, Vesak Day, Hari Raya Haji, Deepavali
+- United Kingdom - Coronation Bank Holiday
 ### Changed
 - France - Holidays revised, county information revised
 ### Deprecated